Migrate merge train PR feedback to FastAPI
